gradio_requirements.txt
gradio>=4.0.0
opencv-python>=4.5.0
numpy>=1.20.0
torch>=1.8.0
torchvision>=0.9.0
PyYAML>=5.4.0
Pillow>=8.0.0 